feat: add pet asset ipc and use in renderer
