+++ src/utils/pet.ts
@@ -1,5 +1,4 @@
-import fs from 'fs'
-import path from 'path'
+
 
 export function weightedRandom<T extends string | number>(chances: Record<T, number>): T {
   const entries = Object.entries(chances) as Array<[T, number]>
@@ -12,47 +11,8 @@ export function weightedRandom<T extends string | number>(chances: Record<T, num
   return entries[entries.length - 1][0]
 }
 
-function pickPetDir(dir: string): string | undefined {
-  if (!fs.existsSync(dir)) return undefined
-  const candidates = fs
-    .readdirSync(dir, { withFileTypes: true })
-    .filter(d => d.isDirectory())
-    .map(d => d.name)
-  if (!candidates.length) return undefined
-  const chosen = candidates[Math.floor(Math.random() * candidates.length)]
-  const gif = path.join(dir, chosen, 'front.gif')
-  if (fs.existsSync(gif)) return gif
-  const png = path.join(dir, chosen, 'front.png')
-  if (fs.existsSync(png)) return png
-  return undefined
-}
-
 export function carregarPet(especie: string, elemento: string) {
-  const base = path.join('Assets', 'Mons')
-  const specieDir = path.join(base, especie)
-  let assetPet = pickPetDir(path.join(specieDir, elemento))
-
-  if (!assetPet) {
-    // procurar qualquer pet da especie ignorando elemento
-    const subdirs = fs
-      .readdirSync(specieDir, { withFileTypes: true })
-      .filter(d => d.isDirectory())
-      .map(d => d.name)
-    for (const sub of subdirs) {
-      assetPet = pickPetDir(path.join(specieDir, sub))
-      if (assetPet) break
-    }
-  }
-
-  const evoMp4 = path.join(base, 'evolution.mp4')
-  const evoGif = path.join(base, 'evolution.gif')
-  const animacaoEvolucao = fs.existsSync(evoMp4)
-    ? evoMp4
-    : fs.existsSync(evoGif)
-      ? evoGif
-      : ''
-
-  return { animacaoEvolucao, assetPet }
+  return window.getPetAssets(especie, elemento)
 }
 
 export interface Stats {
@@ -63,7 +23,7 @@ export interface Stats {
   life: number
 }
 
-export function determinarPetFinal(stats: Stats, elemento: string) {
+export async function determinarPetFinal(stats: Stats, elemento: string) {
   const speciesChances = {
     Fera: 17,
     Reptiloide: 17,
@@ -106,7 +66,7 @@ export function determinarPetFinal(stats: Stats, elemento: string) {
   } as const
 
   const raridade = weightedRandom(rarityChances)
-  const { animacaoEvolucao, assetPet } = carregarPet(especie, elemento)
+  const { animacaoEvolucao, assetPet } = await carregarPet(especie, elemento)
 
   return {
     especie,
